@@ -1,12 +1,11 @@
 package com.jeanbarcellos.project106.exception.handler;
 
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.ExceptionMapper;
-import javax.ws.rs.ext.Provider;
-
 import com.jeanbarcellos.core.dto.ErrorResponse;
 import com.jeanbarcellos.core.exception.ValidationException;
 
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.ext.ExceptionMapper;
+import jakarta.ws.rs.ext.Provider;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
