@@ -2,12 +2,12 @@ package com.jeanbarcellos.project106.mapper;
 
 import java.util.List;
 
-import javax.enterprise.context.ApplicationScoped;
-
 import com.jeanbarcellos.core.MapperBase;
 import com.jeanbarcellos.project106.domain.Person;
 import com.jeanbarcellos.project106.dtos.PersonRequest;
 import com.jeanbarcellos.project106.dtos.PersonResponse;
+
+import jakarta.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
 public class PersonMapper extends MapperBase {
